shop/requirements.txt
django==2.2.17
django-extensions==3.0.9
djangorestframework==3.12.2
